UGC-4794 | Integrate our RESTBase facade
Make the Math extension work with our internal RESTBase facade service.
This requires a few changes for the extension to use a separate base URL
for internal requests and for generating user-facing URLs, as the
default logic does not allow for this without modifying
$wgVirtualRestConfig, which would in turn confuse the VisualEditor
extension. Since we need to do that, also make the extension send check
requests with a JSON payload to simplify the service side handling.

Also clean up after 78521db7bb29eba1dd832100777d7dbe03758a42, as that
logic only pertained to styles set for generated PNG images and seems to
be no longer needed.